feat: frame allocator interface

+++ src/process/mod.rs
@@ -2,7 +2,7 @@ use core::ptr::addr_of_mut;
 
 use crate::mm::PAGE_SIZE;
 
-static mut USER_STACK: [i32; PAGE_SIZE >> 2] = [0; PAGE_SIZE >> 2];
+static mut USER_STACK: [i32; (PAGE_SIZE >> 2) as usize] = [0; (PAGE_SIZE >> 2) as usize];
 
 #[repr(C)]
 struct StackStart {
@@ -12,6 +12,10 @@ struct StackStart {
 
 #[unsafe(export_name = "stack_start")]
 static mut STACK_START: StackStart = StackStart {
-    a: unsafe { addr_of_mut!(USER_STACK).cast::<i32>().add(PAGE_SIZE >> 2) },
+    a: unsafe {
+        addr_of_mut!(USER_STACK)
+            .cast::<i32>()
+            .add((PAGE_SIZE >> 2) as usize)
+    },
     b: 0x10,
 };
